# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EcoStride API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /public/events/{slug}:
    get:
      summary: Get event
      tags: [Public Events]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Not Found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /public/results/events/{slug}:
    get:
      summary: List results
      tags: [Public Results]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
  /admin/events:
    get:
      summary: List events
      tags: [Admin Events]
      responses:
        "200": { description: OK }
    post:
      summary: Create event
      tags: [Admin Events]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "201": { description: Created }
  /admin/events/{id}:
    put:
      summary: Update event
      tags: [Admin Events]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      summary: Delete event
      tags: [Admin Events]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: "snake_case field name"
                    example: "email_address"
                  issue:
                    type: string
                    example: "must be a valid email"
...